p_chara: 97.6269% -> 97.9069% (CColor::operator*(float): implicit u8->float->u8 conversions, no explicit int cast; draw 93.94 -> 96.53)

+++ include/ffcc/color.h
@@ -38,10 +38,10 @@ public:
 	CColor operator*(float scale) const
 	{
 		CColor out;
-		out.color.r = static_cast<unsigned char>(static_cast<int>(color.r * scale));
-		out.color.g = static_cast<unsigned char>(static_cast<int>(color.g * scale));
-		out.color.b = static_cast<unsigned char>(static_cast<int>(color.b * scale));
-		out.color.a = static_cast<unsigned char>(static_cast<int>(color.a * scale));
+		out.color.r = static_cast<unsigned char>(color.r * scale);
+		out.color.g = static_cast<unsigned char>(color.g * scale);
+		out.color.b = static_cast<unsigned char>(color.b * scale);
+		out.color.a = static_cast<unsigned char>(color.a * scale);
 		return out;
 	}
 	/*
